assessment-service/requirements.txt
Flask
langchain
langchain-aws
boto3
python-dotenv
